Sort-of fix command line interface
If getopt() throws an exception while parsing, all options of the
result will be discarded, as definition of options and parsing is
done in one step apparently. This means that if the user fucks up,
he won't be able to automatically get a help message thrown at their
face, which is sort of dumb.

Also, fixed program happily accepting no files at all.